requirements.txt
typer==0.9.0
rich==13.7.0
python-dotenv==1.0.1
openai==1.12.0
httpx==0.27.0
PyGithub==2.3.0
